Code rearrange
